refactor: extract voice control markers into shared voice-control-protocol module
